// frontend/src/generated/http/client/models/ApplyCcmRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { IpaTarget } from './IpaTarget';
export type ApplyCcmRequest = {
    ccm: Array<Array<number>>;
    ct?: number | null;
    target?: (null | IpaTarget);
};