// frontend/src/generated/http/client/models/IdeInfo.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type IdeInfo = {
    artifacts_dir?: string | null;
    bind_addr?: string | null;
    enabled: boolean;
    port?: number | null;
    projects_dir?: string | null;
    sdk_dir?: string | null;
    tools_dir?: string | null;
    url?: string | null;
    workspace_dir?: string | null;
};