// frontend/src/generated/http/client/models/PowerStatusPayload.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PowerStatusPayload = {
    amps?: number | null;
    errors?: any[] | null;
    sources?: any[] | null;
    updated_at?: string | null;
    volts?: number | null;
    watts?: number | null;
};